+++ cloud/blockstore/public/sdk/go/client/error_test.go
@@ -35,27 +35,6 @@ func TestGetClientError(t *testing.T) {
 	require.Equal(t, message, e.Message)
 }
 
-func TestIsRetriable(t *testing.T) {
-	retriable := []uint32{
-		E_REJECTED, E_TIMEOUT, E_THROTTLED, E_OUT_OF_SPACE,
-		E_RDMA_UNAVAILABLE, E_TRANSPORT_ERROR,
-	}
-	for _, code := range retriable {
-		err := &ClientError{Code: code}
-		require.True(
-			t, err.IsRetriable(),
-			"expected retriable: %v", formatErrorCode(code))
-	}
-
-	notRetriable := []uint32{E_ARGUMENT, E_FAIL, E_IO, E_NOT_FOUND}
-	for _, code := range notRetriable {
-		err := &ClientError{Code: code}
-		require.False(
-			t, err.IsRetriable(),
-			"expected not retriable: %v", formatErrorCode(code))
-	}
-}
-
 func TestIsDiskNotFoundError(t *testing.T) {
 	clientErr := &ClientError{}
 	err := clientErr
