+++ cloud/filestore/public/sdk/go/client/error.go
@@ -38,7 +38,8 @@ func (e *ClientError) Error() string {
 
 func (e *ClientError) IsRetriable() bool {
 	switch e.Code {
-	case E_REJECTED, E_TIMEOUT, E_FS_OUT_OF_SPACE, E_FS_THROTTLED:
+	case E_REJECTED, E_TIMEOUT, E_FS_OUT_OF_SPACE, E_FS_THROTTLED,
+		E_TRANSPORT_ERROR:
 		// special error code for retries
 		return true
 	}
