+++ cloud/filestore/public/sdk/go/client/ya.make
@@ -20,6 +20,7 @@ SRCS(
 
 GO_TEST_SRCS(
     durable_test.go
+    error_test.go
     grpc_test.go
 )
